feat: klassementen werken, nu alleen submit functie afmaken
